requirements.txt
alembic==1.17.2
apifairy==1.5.1
apispec==6.8.4
astroid==4.0.2
Authlib==1.6.5
babel==2.17.0
beautifulsoup4==4.14.2
blinker==1.9.0
cffi==2.0.0
click==8.3.1
cryptography==46.0.3
dnspython==2.8.0
eventlet==0.40.3
Flask==3.1.2
faker==33.0.0
flask-cors==6.0.1
Flask-HTTPAuth==4.8.0
flask-marshmallow==1.3.0
Flask-Limiter==3.10.1
Flask-Migrate==4.1.0
Flask-Moment==1.0.6
Flask-SQLAlchemy==3.1.1
future==1.0.0
greenlet==3.2.4
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
json5==0.12.1
Mako==1.3.10
MarkupSafe==3.0.3
marshmallow==4.1.0
marshmallow-sqlalchemy==1.4.2
nltk==3.9.2
numpy==2.3.5
packaging==25.0
psycopg2-binary==2.9.11
pycparser==2.23
pyspellchecker==0.8.3
pytest==8.3.3
pytest-flask==1.3.0
pytest-cov==6.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
regex==2025.11.3
scikit-learn==1.7.2
scipy==1.16.3
six==1.17.0
soupsieve==2.8
SQLAlchemy==2.0.44
threadpoolctl==3.6.0
tqdm==4.67.1
typing_extensions==4.15.0
urllib3==2.5.0
webargs==8.7.1
Werkzeug==3.1.3
wrapt==2.0.1
